Move files from `/tests_models` to `/tests`.
