--- backends/nxp/tests_models/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2026 NXP
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
